feat: save movie requests in global context
